requirements.txt
mistralai==0.0.12
python-dotenv==1.0.0
Flask==3.0.0
